remove configurtion from component to a seprate file according to best practices
