+++ src/config/index.js
@@ -1,0 +1,11 @@
+import * as firebase from 'firebase';
+
+
+var config = {
+    apiKey: "AIzaSyBwHrtA3ER0d4V9JoIw3oBbwaYs2NiGMhg",
+    authDomain: "project-level-1.firebaseapp.com",
+    databaseURL: "https://project-level-1.firebaseio.com",
+    storageBucket: "project-level-1.appspot.com",
+    messagingSenderId: "835802635181"
+  };
+  firebase.initializeApp(config);
